project/requirements.txt
fastapi==0.70.1
uvicorn==0.16.0
gunicorn==20.1.0
asyncpg==0.25.0
tortoise-orm==0.17.8
aerich==0.6.1
requests==2.26.0
pytest==6.2.5
pytest-cov==3.0.0
pytest-xdist==2.5.0
flake8==4.0.1
black==21.12b0
isort==5.10.1
newspaper3k==0.2.8